smartcare_api/serializers.py
from rest_framework import serializers
from appointments.models import Appointment
from authentication.models import User

class AppointmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Appointment
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'user_type']