requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
dj-database-url==0.5.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0